acme_project/requirements.txt
asgiref==3.8.1
Django==5.2.3
django-bootstrap5==24.2
django_debug_toolbar==3.8.1
pillow==11.2.1
pytz==2025.2
sqlparse==0.5.3
tzdata==2025.2
